+++ structurizr-dsl/src/main/java/com/structurizr/dsl/StructurizrDslExpressions.java
@@ -3,6 +3,8 @@ package com.structurizr.dsl;
 class StructurizrDslExpressions {
 
     static final String ELEMENT_TYPE_EQUALS_EXPRESSION = "element.type==";
+    static final String ELEMENT_TECHNOLOGY_EQUALS_EXPRESSION = "element.technology==";
+    static final String ELEMENT_TECHNOLOGY_NOT_EQUALS_EXPRESSION = "element.technology!=";
     static final String ELEMENT_TAG_EQUALS_EXPRESSION = "element.tag==";
     static final String ELEMENT_TAG_NOT_EQUALS_EXPRESSION = "element.tag!=";
     static final String ELEMENT_PROPERTY_EQUALS_EXPRESSION = "element\\.properties\\[.*]==.*";